main

+++ 코드/Model/NODE.py
@@ -17,7 +17,7 @@ class RNNEncoder(nn.Module):
         # t = t.clone()
         t[1:] = t[:-1] - t[1:]
         t[0] = 0.
-        xt = torch.cat((x, t), dim=-1)
+        xt = torch.cat((x, t.type(torch.float32)), dim=-1)
 
         _, h0 = self.rnn(xt.flip((0,)))  # Reversed
         # Compute latent dimension
@@ -63,6 +63,7 @@ class NODE(nn.Module):
 
     def forward(self, x, t, MAP=False):
         z_mean, z_log_var= self.encoder(x, t[:-1])
+
         if MAP:
             z = z_mean
         else:
